simulation_tool/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
numpy>=1.20.0
pydantic>=2.0.0
